more changes from the Windows port
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@425 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/d/types.h
@@ -108,7 +108,7 @@ extern char *libfac_version;
 #endif
 
 #ifndef PAGESIZE
-extern int getpagesize();
+extern size_t getpagesize();
 #define PAGESIZE getpagesize()
 #endif
 #define RUP(x) ((((x) + PAGESIZE - 1) / PAGESIZE) * PAGESIZE)
